Date ranges filter
Added support for date ranges.
There is a field that stores a date for sessions but is stored as a string and the format is not standardized, this will be updated during future versions and implemented a fix in the next api endpoint version to allow for a smoother transition.